Minor refactorings and improvements

+++ src/TodoApp.Application/Todos/Commands/Delete/DeleteTodoCommandHandler.cs
@@ -19,7 +19,7 @@ public class DeleteTodoCommandHandler : IRequestHandler<DeleteTodoCommand, Unit>
 
 	public async Task<Unit> Handle(DeleteTodoCommand command, CancellationToken cancellationToken)
 	{
-		if (await _userRepository.GetUserById(command.UserId) is null)
+		if (await _userRepository.GetUserById(command.UserId) is not User user)
 		{
 			throw new ApiException("Not Found", "User does not exist.", HttpStatusCode.NotFound);
 		}
@@ -28,6 +28,13 @@ public class DeleteTodoCommandHandler : IRequestHandler<DeleteTodoCommand, Unit>
 		{
 			throw new ApiException("Not Found", "Todo does not exist.", HttpStatusCode.NotFound);
 		}
+		
+		if (todo.UserId != user.Id)
+		{
+			throw new ApiException("Unauthorized", "You do not have access to this resource.",
+				HttpStatusCode.Unauthorized);
+		}
+
 
 		await _todoRepository.DeleteTodo(todo.Id);
